push_swap is dead

+++ libft/libft.h
@@ -25,7 +25,7 @@ typedef struct s_list
 }	t_list;
 
 int		ft_isprint(int c);
-int		ft_isdigit(int c);
+long	ft_isdigit(long c);
 int		ft_isascii(int c);
 int		ft_isalpha(int c);
 int		ft_isalnum(int c);
